+++ app/models/restaurant.py
@@ -56,11 +56,11 @@ class Restaurant(db.Model):
                                 "booked": obj.booked}
                                 for obj in self.reservation],
             'reviews': {obj.id: {"rating": obj.rating,
-                                "comment": obj.comment,
-                                "restaurantId": obj.restaurantId,
-                                "userId": obj.userId,
-                                "id": obj.id}
-                                for obj in self.restaurant_review},
+                                 "comment": obj.comment,
+                                 "restaurantId": obj.restaurantId,
+                                 "userId": obj.userId,
+                                 "id": obj.id}
+                        for obj in self.restaurant_review},
             'cover_photo': self.cover_photo,
             'cuisine_type': type,
             'user_id':self.user_id
